style: вынесены типы статуса практики в shared
Типы используется не только в модуле практики
